refactor: handler to service-controller - table and column
