+++ packages/nocodb/src/lib/services/index.ts
@@ -1,2 +1,4 @@
 // export * as projectService from './projectService';
 export * as tableService from './tableService';
+export * as columnService from './columnService';
+export * as filterService from './filterService';
